feat: more robust CLCS handling
One of the most common issues in the integration tests are projection
errors when cartesian coordinates are converted to curvilinear coordinates.
To address those issues, the handling of CLCS is streamlined to consistently
use a large step fallback CLCS and use consistent pre-processing for
the reference paths of the CLCSs.