feat(common): automatically compute lanelet assignments for vehicles
Currently, vehicles can only be constructed if an explict lanelet assignment is provided.
This is not very ergonomic and also makes each vehicle static, since originally
its lanelet assignment were not expected to change.

This adds the option to automatically compute the lanelet assignments for
vehicles on demand, without any significant change to the API.